api/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html                                                                                                                                                             
torch
torchivision
numpy
pandas
matplotlib
pandas

torch
torchvision
numpy
flask
seaborn
gunicorn
flask-cors
pyfcm
python-decouple